rename dialects package to parsers
